requirements.txt
fastapi
uvicorn
sqlite-vec
python-multipart
Pillow
pydantic
python-json-logger
torch
transformers
openai
jinja2
aiofiles
numpy
requests
scikit-learn
google-genai
